MS-38: refactor all the layout from signup
